Improve performance when writing lp and reading sol files
We improve the performance of the solution file parser in the case of
cplex, by making use of the more performant quick-xml crate and
buffering the reads through `BufReader`. quick-xml indeed requires the
reader to be buffered, probably to avoid making too many read syscalls.

Buffering is also applied to writing the lp file. Manual benchmarks are
showing significant improvements over non-buffered read/writes for
problem sizes of several thousands of variables